zawiesza sie wprowadzam kolejne zmiany w plikach likwidujac pojawiajace sie bledy podczas ksiegowania

+++ npkpir_23/src/java/dao/DAO.java
@@ -78,7 +78,9 @@ public abstract class DAO<T> {
     }
     
     public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+        if (entity!=null) {
+            getEntityManager().remove(getEntityManager().merge(entity));
+        }
     }
     
     public void removeList(List<T> entityList) {
